+++ lib/src/ops/instance_norm_op.dart
@@ -106,7 +106,7 @@ class InstanceNormOp extends TransformOp
       );
     }
 
-    if (eps <= 0) {
+    if (!eps.isFinite || eps <= 0) {
       throw InvalidParameterException(
         'eps',
         eps.toString(),