+++ lib/src/ops/layer_norm_op.dart
@@ -152,7 +152,7 @@ class LayerNormOp extends TransformOp
       );
     }
 
-    if (eps <= 0) {
+    if (!eps.isFinite || eps <= 0) {
       throw InvalidParameterException(
         'eps',
         eps.toString(),