+++ lib/src/ops/rms_norm_op.dart
@@ -128,7 +128,7 @@ class RMSNormOp extends TransformOp with InPlaceTransform, RequiresContiguous {
       );
     }
 
-    if (eps <= 0) {
+    if (!eps.isFinite || eps <= 0) {
       throw InvalidParameterException(
         'eps',
         eps.toString(),